requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
requests==2.32.3
python-dotenv==1.0.1
jinja2==3.1.5
pytest==8.3.4

